[ci skip] Added the timeout back for the transport.
I was running into a problem where i would constantly get an io/timeout
error when trying to read the RCP message from the leader. I cannot
for the life of me reproduce this issue in any other net code I've
used or written. And I believe the problem lies in hashicorps encoding
and decoding library that they wrote. The library basically takes any
object and turns it into bytes that are typesafe when being decoded.

But I think the way they are reading from the connection is lazy enough
and the serialization is slow enough that we are hitting timeout walls
or its locking up completely somewhere and causing a timeout because the
net.Conn gets left alone for too long with nothing being read or written.

I think the best solution would be to implement a better encoding and
decoding library because I'm still sometimes hitting an io/timeout with
a 30 second constraint.

The io/timeout error is consistently thrown in the decodeRpc
function in net_transport.go. When the first decode is being attempted.

I think a better encoding and decoding solution would be to implement a
true to PostgreSQL protocol, which is a single byte prefix for the
message type, then a 4 byte length for the message itself.

This would allow the reader to grab the entire message from the wire
and potentially issue a simple "I received your message" response while
it processes the actual message. That might not even be necessary though
if the problem truly does lie in hashicorps serialization library. A
very deliberate encode and decode method for each message type would be
significantly faster than ANY reflection implementation and might just
fix the issue entirely.