[ci skip] Working on getting new/custom transport built out.
This new transport will include the new net code I mentioned in my
previous commit. This net code should be significantly faster than
using message pack or whatever raft was using before. This is purely
because the encoding and decoding will be hard coded for each message
type and no reflection is needed at all. This should also reduce the
size of messages being sent. For example an AppendEntries heartbeat
would only be the following. (number represents the number of bytes)

1 Message type prefix
4 Message length
4 raft protocol version
8 raft term
4 leader length
>= 0 leader bytes
8 previous log entry
8 previous log term
2 number of entries
-- for each entry
4 size of entry
-- if entry size is -1, entry is null
8 entry index
8 entry term
1 entry log type
4 entry data length
>= 0 entry data

All in all, for messages like heartbeats this new encoding would make
extremely small messages to be sent over the wire, and will be decoded
very very quickly.