[ci skip] This is a messy commit.
Basically right now the raft cluster is the most stable it's ever been.
With default raft configuration nodes can join and leave the cluster
VERY fluidly and I'm no longer seeing any io/timeout errors. And this
is with my new transport which is still respecting all of the timeouts
that raft wants.

This code will need to be cleaned up a bit, but this is a pretty good
marker for successfully building out a new transport.