Minor updates, hopefully a good ci build too.
I've just tested the new net code with noahdb running in kubernetes.
This is where we were seeing the most io/timeouts. I'm not sure if that
is due to minikube, or if it's that and a combination of bad net code?

But currently running 5 nodes in minikube for several minutes, they
instantly joined together to form a cluster automatically using the
discovery code. And as soon as a leader was elected they became
completely stable, no errors in the logs at all, and all state changes
are being replicated smoothly. And all of that with the default
heartbeat timeout of 1 second.