Fixed data race in raft backend.
This was due to returning a pointer to a variable that we store on
the backend struct itself. This code was pretty much copied from go/pgx
and maybe that library is not vulnerable to it because of the flow of
data and lack of concurrency. But due to the way that raft seems to work
and its high level of concurrency; returning a pointer to a variable
that might then be overwritten by a subsequent message makes for a bad
time.

This resolves the issue by declaring a new variable and returning that
rather than pointing to an existing one. This is probably less memory
efficient than the original way, but will fix the race condition.