More prepared statements work.
Previously I was trying to infer the prepared statements param type by
looking at the columns, and then assuming based on the column or it's
typecase what type of param it was. But this was incorrect and would
not include any params that were in a where clause or in a join clause.

This includes some style changes in the sql package and in the queryutil
package a new function called GetArgumentsEx which will help derive
types for parameters.