Adding more tests around the SELECT planner. Improved ID resolution.
The SELECT query planner is not much better at finding IDs in the query,
and it can now even handle strings passed as IDs as long as they can be
converted to a uint64.

However the planner is currently assuming an IN or EQUAL comparision
when looking at tenant IDs. Meaning if you did something like

WHERE tenant_id != 5

It would treat that as

WHERE tenant_id = 5

Since it is always assuming an inclusive comparision at this time.