Finished implementing basic handling of prepared statements.
When a prepared statement is being executed, the runner will transition
to an extended query mode. When in this mode all of the queries executed
will be sent to the data node as extended queries rather than being
sent as simple queries, this also allows it to properly tunnel the
response back to the client rather than needing to process the response
manually.