Fixed error in regression tests. The issue was with prepared statements.
When a prepared statement is executed, noahdb would send the extended
query protocol messages to the data node and prepare the statement with
the name "1". This would cause problems when the connection would be
released back into the pool and used again by the following test where
the same prepared statement name would try to be used but would cause
an error because it already existed. This fixes that by not naming the
prepared statements on the data node at all.