Bumped timber version. Starting to think about transactions. + more.
Starting to make changes to how connections are allocated per query
session. Right now a connection is allocated per query issued to a data
node. Once I start to add support for transactions, connections will
need to persist past the life of a single query and continue for the
length of the transaction. But can be born at any point during a
transaction depending on what shards are hit throughout.